add optional DNS proxy
this is to support deployment scenarios wherein the worker is run with
/etc/resolv.conf containing a `127.x.x.x` address. this will not work
from within Garden's containers as they run in their own network
namespace.

if enabled, a DNS server is started concurrently with Garden, and Garden
is configured with it as an additional DNS server for its containers.
the DNS server then proxies to the servers configured in
/etc/resolv.conf, e.g. 127.0.0.11 in a Docker container.

concourse/concourse#1145